requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.3.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
distro==1.9.0
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
flask-restx==1.3.0
fsspec==2024.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openai==1.30.2
packaging==24.0
pillow==10.3.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-whois==0.9.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
rpds-py==0.18.1
safetensors==0.4.3
six==1.16.0
sniffio==1.3.1
sympy==1.12
tokenizers==0.19.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.4
transformers @ git+https://github.com/huggingface/transformers@42d8dd8716d45291733191464e249a83b93e4d86
typing_extensions==4.12.0
urllib3==2.2.1
Werkzeug==3.0.3
